bugfix: Fixed passing in job names to certain functions without environment prefix
Functions to determine if job exists or to tie trigger to job were using the name
of the job that did not have the environment prefix attached. This commit corrects
that and also updates the readme to reference the poetry project file instead of the
setup file with regards to the version number.